Added throttle to SingleStageCraft -Added new function set_throttle_by_name. Does what it says. -Modified toggle_engines_by_isp and turn_engines_off to use this interface. -Removed any functionality of engine_activations
As of now this shouldn't be much, if any, of a performance hit.

Changed simulate_landing_with_craft to micromanage throttle
-Also changed the logging output to give engine name and throttle instead of isp

Fixed a bug where thrust effectiveness wouldn't get updated immediately after changing throttle.
This should have minimal effects on calculations since it only affects one physics tick after an
engine gets turned on, but it was still an inaccuracy.